refactor: Improve performance of GetTVShowsByActor
This commit refactors the GetTVShowsByActor function to improve its performance. The startIndex and endIndex variables are now casted to int, which fixes a bug that caused the function to return incorrect results. Additionally, the code now uses a slice instead of an array for extractedTVShows and lockIndexes, which reduces memory usage.

+++ tmdb/tmdb.go
@@ -476,11 +476,11 @@ func (m *mediaClient) GetTVShowsByActor(actorID int, page int) (*PaginatedTVShow
 		return nil, err
 	}
 	var wg sync.WaitGroup
-	var startIndex = math.Min(float64((page-1)*20), float64(len(actorTVCredits.Cast)-1))
-	var endIndex = math.Min(float64(page*20), float64(len(actorTVCredits.Cast)))
+	var startIndex = int(math.Min(float64((page-1)*20), float64(len(actorTVCredits.Cast)-1)))
+	var endIndex = int(math.Min(float64(page*20), float64(len(actorTVCredits.Cast))))
 	var extractedTVShows = make([]*TVShow, endIndex-startIndex)
 	var lockIndexes = make([]sync.Mutex, endIndex-startIndex)
-	for index, tvShow := range actorTVCredits.Cast[int(startIndex):int(endIndex)] {
+	for index, tvShow := range actorTVCredits.Cast[startIndex:endIndex] {
 		wg.Add(1)
 		go func(tvShowID, index int) {
 			defer wg.Done()
